src/test/kotlin/day24/LobbyLayoutTest.kt
package day24

import org.assertj.core.api.Assertions.assertThat
import org.junit.jupiter.api.Test

internal class LobbyLayoutTest {

    private val input = """
        sesenwnenenewseeswwswswwnenewsewsw
        neeenesenwnwwswnenewnwwsewnenwseswesw
        seswneswswsenwwnwse
        nwnwneseeswswnenewneswwnewseswneseene
        swweswneswnenwsewnwneneseenw
        eesenwseswswnenwswnwnwsewwnwsene
        sewnenenenesenwsewnenwwwse
        wenwwweseeeweswwwnwwe
        wsweesenenewnwwnwsenewsenwwsesesenwne
        neeswseenwwswnwswswnw
        nenwswwsewswnenenewsenwsenwnesesenew
        enewnwewneswsewnwswenweswnenwsenwsw
        sweneswneswneneenwnewenewwneswswnese
        swwesenesewenwneswnwwneseswwne
        enesenwswwswneneswsenwnewswseenwsese
        wnwnesenesenenwwnenwsewesewsesesew
        nenewswnwewswnenesenwnesewesw
        eneswnwswnwsenenwnwnwwseeswneewsenese
        neswnwewnwnwseenwseesewsenwsweewe
        wseweeenwnesenwwwswnew
    """.trimIndent().lines()

    @Test
    fun `should count black tiles`() {
        assertThat(LobbyLayout().findBlackTiles(input)).isEqualTo(10)
    }

    @Test
    fun `should count black tiles after 100 days`() {
        assertThat(LobbyLayout().findBlackTilesAfter100Days(input)).isEqualTo(2208)
    }
}